splyttr-api/requirements.txt
coreapi==1.32.0
defusedxml==0.4.1
Django==1.10
django-allauth==0.26.1
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-filter==0.13.0
django-rest-auth==0.8.1
django-rest-swagger==2.0.3
djangorestframework==3.4.1
gunicorn==19.6.0
itypes==1.1.0
Markdown==2.6.6
oauthlib==1.1.2
openapi-codec==1.0.0
psycopg2==2.6.2
python3-openid==3.0.10
PyYAML==3.11
requests==2.10.0
requests-oauthlib==0.6.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
